:recycle: Separate functions about status
